+++ src/main/org/codehaus/groovy/runtime/dgmimpl/NumberNumberMultiply.java
@@ -170,7 +170,7 @@ public final class NumberNumberMultiply extends NumberNumberMetaMethod {
         public final Object callBinop(Object receiver, Object arg) {
             try {
                 if (checkPojoMetaClass()) {
-                    return new Double(((Double) receiver).doubleValue() * (double)((Integer) arg).intValue());
+                    return new Double(((Double) receiver).doubleValue() * ((Integer) arg).intValue());
                 }
             }
             catch (ClassCastException e) {//
@@ -183,7 +183,7 @@ public final class NumberNumberMultiply extends NumberNumberMetaMethod {
         }
 
         public final Object invokeBinop(Object receiver, Object arg) {
-            return new Double(((Double) receiver).doubleValue() * (double)((Integer) arg).intValue());
+            return new Double(((Double) receiver).doubleValue() * ((Integer) arg).intValue());
         }
     }
