+++ src/main/org/codehaus/groovy/runtime/dgmimpl/NumberNumberPlus.java
@@ -191,7 +191,7 @@ public final class NumberNumberPlus extends NumberNumberMetaMethod {
         public final Object callBinop(Object receiver, Object arg) {
             try {
                 if (checkPojoMetaClass()) {
-                    return new Double(((Double) receiver).doubleValue() + (double)((Integer) arg).intValue());
+                    return new Double(((Double) receiver).doubleValue() + ((Integer) arg).intValue());
                 }
             }
             catch (ClassCastException e) {//
@@ -204,7 +204,7 @@ public final class NumberNumberPlus extends NumberNumberMetaMethod {
         }
 
         public final Object invokeBinop(Object receiver, Object arg) {
-            return new Double(((Double) receiver).doubleValue() + (double)((Integer) arg).intValue());
+            return new Double(((Double) receiver).doubleValue() + ((Integer) arg).intValue());
         }
     }
